src/backend/requirements.txt
openai
pandas
matplotlib
seaborn
pickle5
numpy
docstring_parser
langchain
colorama
scikit-learn
lxml
parse
comment-parser
beautifulsoup4
replicate
pyopenssl
llama-index-vector-stores-qdrant